STSMACOM-267 Created separate CustomFieldsDragDropContext component
