+++ lib/CustomFields/components/CustomFieldsForm/CustomFieldsForm.js
@@ -27,6 +27,7 @@ import {
   AddButton,
   DeleteModal,
   CustomFieldsAccordions,
+  CustomFieldsDragDropContext,
 } from './components';
 
 import {
@@ -271,16 +272,20 @@ const CustomFieldsForm = ({
                 <>
                   {!!fields.length && renderExpandAllButton()}
                   {!!fields.length && (
-                    <CustomFieldsAccordions
+                    <CustomFieldsDragDropContext
                       fields={fields}
-                      permissions={permissions}
-                      accordions={accordions}
-                      onToggleAccordion={onToggleAccordion}
                       onDragEnd={onDragEnd}
-                      onDeleteClick={onDeleteClick}
-                      changeFieldValue={form.change}
-                      getFormState={form.getState}
-                    />
+                    >
+                      <CustomFieldsAccordions
+                        fields={fields}
+                        permissions={permissions}
+                        accordions={accordions}
+                        onToggleAccordion={onToggleAccordion}
+                        onDeleteClick={onDeleteClick}
+                        changeFieldValue={form.change}
+                        getFormState={form.getState}
+                      />
+                    </CustomFieldsDragDropContext>
                   )}
                   <AddButton
                     options={getFieldTypesOptions()}
