+++ lib/CustomFields/components/CustomFieldsForm/components/CustomFieldsDragDropContext/index.js
@@ -1,0 +1,1 @@
+export { default } from './CustomFieldsDragDropContext';
