+++ lib/CustomFields/components/CustomFieldsForm/components/index.js
@@ -1,3 +1,4 @@
 export { default as AddButton } from './AddButton';
 export { default as DeleteModal } from './DeleteModal';
 export { default as CustomFieldsAccordions } from './CustomFieldsAccordions';
+export { default as CustomFieldsDragDropContext } from './CustomFieldsDragDropContext';
